feat(admin): dual-list Compendium review and close AFMPS P0-3 docs
Improve Compendium import control (À contrôler / Prêtes / Exclues, filters,
pagination) and align pharmacy roadmap docs with staging+prod AFMPS catalogue.
Also harden import-cnk for app-role gates and AFMPS size checks as int64.

+++ go/internal/handlers/afmps_import.go
@@ -69,7 +69,7 @@ func (a *API) adminCreateAFMPSImport(w http.ResponseWriter, r *http.Request) {
 		writeErr(w, r, http.StatusBadRequest, "bad_request", "empty_file")
 		return
 	}
-	if len(data) > pharmacy.MaxAFMPSCSVBytes {
+	if int64(len(data)) > pharmacy.MaxAFMPSCSVBytes {
 		writeErr(w, r, http.StatusBadRequest, "bad_request", "file_too_large")
 		return
 	}
@@ -306,7 +306,7 @@ func (a *API) internalAfmpsImportRun(w http.ResponseWriter, r *http.Request) {
 		})
 		return
 	}
-	if len(data) > pharmacy.MaxAFMPSCSVBytes {
+	if int64(len(data)) > pharmacy.MaxAFMPSCSVBytes {
 		a.notifyAfmpsImportOps(r.Context(), "file_too_large",
 			fmt.Sprintf("CSV AFMPS trop volumineux (>%d MiB) : %s", pharmacy.MaxAFMPSCSVBytes>>20, objectKey),
 			"/admin/afmps-imports", "")
